Width change in hero section from 100vw to 100%
This has helped avoid catching the scrollbar on
the right side of the screen.

About section now responsive. There is a small
glitch in the height of the first div containing
the vision statement in screens smaller than
1024px. The height creates a visible difference
in white space between that div and the others.
Probably needs some further looking into
at a later time.